Backend/requirements.txt
google-generativeai
requests==2.31.0
Pillow==10.4.0
fastapi==0.100.0
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.11.4
reportlab